Update for EFI 2021
Added Support for HLA data files already present in the prepareData
directory.

Update Text for getting started

Adapted detailedExplanationPrepareData.doc to include a "Semi-Automated
Way" for data preparation explaining that users can download inputfiles
manually and then use the BuildData script.

Hint for Semi-Automated Data preparion in Word file for Windows

Troubleshooting broken BuilData input files

Added a section to detailedExplanationPrepareData.doc to hint on removal
and replacement of potentially broken (that is, e.g. incomplete or corrupted)
inputfiles for BuildData.py

Updated Detailed getting started Windows

Added a section about BuildData and input files that hints on the
semi-automated way to manually download the input files.

Trimming Input file strings

Removed a bug that included trailing spaces in the inpute files after
eg. a filename as the filename.

Set seed to zero

Added std∷ to make unique

User Guide Updates

Rolled out updates to user guides to PDF files and removed the review
version in doc files.

+++ prepareData/DownloadDataP2.py
@@ -30,39 +30,74 @@
 
 
 import urllib2
-import zipfile 
+import zipfile
+import os.path
+import sys
 
 def downloadAndExtractData():
 
     print('Download data files')
 
-    print('    Download hla_nom_p.txt')
-    with open('hla_nom_p.txt', 'wb') as f:
-        f.write(urllib2.urlopen('http://hla.alleles.org/wmda/hla_nom_p.txt').read())
-        f.close()
+    successStatus = 0
 
-    print('    Download hla_nom_g.txt')
-    with open('hla_nom_g.txt', 'wb') as f:
-        f.write(urllib2.urlopen('http://hla.alleles.org/wmda/hla_nom_g.txt').read())
-        f.close()
+    try:
 
-    print('    Download alpha.v3.zip')
-    with open('alpha.v3.zip','wb') as f:
-        f.write(urllib2.urlopen('https://bioinformatics.bethematchclinical.org/HLA/alpha.v3.zip').read())
-        f.close()
-    print('    Extract alpha.v3.zip')
-    alphav3Zip = zipfile.ZipFile('alpha.v3.zip')
-    alphav3Zip.extract('alpha.v3.txt')
-
-    print('    Download hla_ambigs_xml.zip')
-    with open('hla_ambigs.xml.zip','wb') as f:
-        f.write(urllib2.urlopen('https://github.com/jrob119/IMGTHLA/raw/Latest/xml/hla_ambigs.xml.zip').read())
-        f.close()
-    print('    Extract hla_ambigs_xml.zip')
-    hlaAmbigZip = zipfile.ZipFile('hla_ambigs.xml.zip')
-    hlaAmbigZip.extract('hla_ambigs.xml')
+        if os.path.exists('hla_nom_p.txt'):
+            print('    File hla_nom_p.txt already present in directory.')
+        else:
+            print('    Download hla_nom_p.txt')
+            with open('hla_nom_p.txt', 'wb') as f:
+                f.write(urllib2.urlopen('http://hla.alleles.org/wmda/hla_nom_p.txt').read())
+                f.close()
 
 
+        if os.path.exists('hla_nom_g.txt'):
+            print('    File hla_nom_g.txt already present in directory.')
+        else:
+            print('    Download hla_nom_g.txt')
+            with open('hla_nom_g.txt', 'wb') as f:
+                f.write(urllib2.urlopen('http://hla.alleles.org/wmda/hla_nom_g.txt').read())
+                f.close()
+
+
+        if os.path.exists('alpha.v3.zip'):
+            print('    File alpha.v3.zip already present in directory.')
+        else:
+            print('    Download alpha.v3.zip')
+            with open('alpha.v3.zip','wb') as f:
+                f.write(urllib2.urlopen('https://bioinformatics.bethematchclinical.org/HLA/alpha.v3.zip').read())
+                f.close()
+
+        
+        if os.path.exists('hla_ambigs.xml.zip'):
+            print('    File hla_ambigs.xml.zip already present in directory.')
+        else:
+            print('    Download hla_ambigs_xml.zip')
+            with open('hla_ambigs.xml.zip','wb') as f:
+                f.write(urllib2.urlopen('https://github.com/jrob119/IMGTHLA/raw/Latest/xml/hla_ambigs.xml.zip').read())
+                f.close()
+
+
+    except: 
+        sys.stderr.write('Something went wrong downloading files. You might want to check proxies or firewalls. You can also manually download the files.')
+        successStatus = 1
+
+
+    if successStatus < 1:
+        try:
+            print('    Extract alpha.v3.zip')
+            alphav3Zip = zipfile.ZipFile('alpha.v3.zip')
+            alphav3Zip.extract('alpha.v3.txt')
+            print('    Extract hla_ambigs_xml.zip')
+            hlaAmbigZip = zipfile.ZipFile('hla_ambigs.xml.zip')
+            hlaAmbigZip.extract('hla_ambigs.xml')
+
+        except:
+            sys.stderr.write('Something went wrong extracting files.')
+            successStatus = 2
+
+
+    return successStatus
 
 if __name__ == "__main__":
